refactor: extract AX12, Response, DynamixelFatalError into own modules
Enforces the one-class-per-file rule from AGENTS.md by moving:
- AX12 -> pydynamixel/ax12.py (uses lazy import to avoid circular deps)
- Response -> pydynamixel/response.py (dataclass for decoded status packets)
- DynamixelFatalError -> pydynamixel/dynamixel_fatal_error.py

dynamixel.py now imports these types from their own modules.
__init__.py updated to re-export AX12 from pydynamixel.ax12.
Shebang and encoding headers added to all affected files.
