+++ pydynamixel/__init__.py
@@ -1,6 +1,10 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
 """PyDynamixel package."""
 
 from . import chain, dynamixel, packets, registers
-from .dynamixel import AX12
+from .ax12 import AX12
 
 __all__ = ["AX12", "chain", "dynamixel", "packets", "registers"]
+
